+++ packages/js-toolkit/helpers/index.ts
@@ -8,6 +8,7 @@ export * from './importOnMediaQuery.js';
 export * from './importWhenIdle.js';
 export * from './importWhenPrefersMotion.js';
 export * from './importWhenVisible.js';
+export * from './logTree.js';
 export * from './registerComponent.js';
 export {
   type QueryOptions,